Converted all nav code to FazStacheItem.
Nav looks like the NavBarNav right now with little differences,
like handling tabs. We're positioning to merge the two codebases
for better code reuse and maintenance.

Children list was renamed to items and tabContents to tabs. Before
the attribute tabs was a boolean indicating the nav has tabs, the
attribute was named to hasTabs.

Nav being processed from a json data is working fine but not
handling tabs. We will cover this later.

Fixes: #89 #93