Added react dependencies to the project.
Also added the babel preset react making webpack able to handle
the jsx sintax in js and jsx files. This is being executed by
the babel-loader and presets are present in the webpack.config.js.

The lodash was added also to give us more flexibility while
dealing with arrays and objects. We're losing lots of features
after the removal of canjs and lodash will mitigate that, but
the ways things were handled will change a little bit.

Refs: #110 #112