Added defineProps method to the react item.
Also added attributesToStates to the element item.

The defineProps methos is created to react item decendants finish
defining props instead of implementing the constructor.

We do that because if an element item triggered the react item
attributesToStates method will be called inside react item's
constructor.

The attributesToStates is where the element item handle setting
states to it's react item. To avoid braking react item's
consturctor defineProps should be used to customize props
management.

Refs: #110