Finished faz pagination rendering.
We added from the previous canjs version the block buttons to
speed up navigation.

Buttons will be displayed according the context. For instance, if
current page is the first one, first page button won't be
displayed, if current page is the last one, last page button wont't
be displayed. We can assume the same case for block buttons and for
page buttons, if the pagination isn't multi block for instance.

Also implemented the init callback.

Refs: #113