Implemented a search to find faz element items inside an item.
This will go though the item html child nodes and look for any
faz element item and recursively through their child nodes until
we find an faz element item or we have no child nodes to iterate.

As a depth tree search, using a brute force approach I think it
would be great to implemente a depth limit in order to don't loose
performance.

This depth limit could also indicate that an faz element item in a
5 or 6 depth distance from the faz element container won't be
considered a child and we won't set a child/parent relationship.

Refs: #110