Added method attributesToProps to the element item.
Also positioned the react element to process it's metadata instead
of accumulating solved react elemetns on the fly.

I think it is better too keep meta elements(js arrays/objects)
instead of trying to figure out how to handle/sync the original
dom and reac dom. It seems like that's the right approach pointed
in the react documentation.

The meta elements are also a way to solve json data comming from
a http response and make the object resolve the data the same
way we do while mounting a strucutre derived from webcomponents.

For rendering hierarchical components like navs or navbars this
is the approach we're gonna take by now.

Refs: #110