Added partial implementation of the nav in react.
Changed the json from the examples using the parameter source. The
source example 2 it is a multi-level menu.

Implemented on mouse over and leave timeout triggers to clean
selected submenus. Right now we disabling the current submenu
selected and setting active the previous non-dropdown menu item.
This could be used as a type strategy by configuration but I don't
think this will be the default behavior.

Resfs: #126