refactor(item): event types dispatched will to be all lowercase
After dispaching a property change we call the on{property} method.

For instance:

When this.extraClasses is changed we dispatch the event type
extraclasseschanged and call this.onExtraClassesChanged method.

Listening to the disapached event is better if you are creating a
new component, overriding the on{property} method is better for
the end user or when you are applying the component directly to
your page.

Refs: #112